OMP_NUM_THREADS expect int in MXNet benchmark.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=273383734

+++ perfkitbenchmarker/linux_benchmarks/mxnet_benchmark.py
@@ -302,7 +302,7 @@ def Run(benchmark_spec):
       # Specifies the number of threads to use in CPU test.
       # https://mxnet.incubator.apache.org/faq/perf.html
       mx_benchmark_cmd = 'OMP_NUM_THREADS={omp_num_threads} {cmd}'.format(
-          omp_num_threads=vm.NumCpusForBenchmark() / 2,
+          omp_num_threads=int(vm.NumCpusForBenchmark() / 2),
           cmd=mx_benchmark_cmd)
 
     if num_layers:
